cloud: Add instance spec argument to CreateInstance
Update the CreateInstance method signature to include instance spec.
This is the VM spec that is made available to the provider specific
implementations to enable flexible VM instance type selection

Fixes: #321
